hal: check to avoid metadata update from framework if phone mode is in call
If there is concurrency of gaming/media with voice call, framework will send
metadata update when ringtone stream will end. As media/gaming stream is
active at that instance, it will trigger recong in BT stack. To avoid this
introduced a condition which will not send metadata update to BT if phone
mode is IN_CALL.

Change-Id: Ida92af47b6ef7d01580c90bdc3d97d650ebf8809