HAL: Custom Ringtone Signal Feature Enablement
HAL changes to support implementation of new Custome Ringtone Signal (CRS) feature.
CRS is received from far end device, and a new audio patch is created to play the
CRS audio on near end device while in ringtone mode. CRS is then torn down when
call is answered, rejected, or muted as with a standard ringtone.

Change-Id: I2cbf463bcecf45749bf233fa68b36638eb595c2f