HAL: Enable proper setting of CRS volume in speaker only mode
CRS volume level is properly set when a combo device is received,
however the volume also needs to be properly set when speaker is
the only device. So SetVoiceVolume for CRS needs to be called in
VoiceStart to ensure volume is properly set for speaker only. Also
fixed a channel mismatch issue where loopback device was setup with
one audio channel instead of two. 

Change-Id: Ib740d9561b91f8ca14476267ad0457ae173b27e5