HAL: Add support for CRS initialization from voice setMode and setParam
The CRS usecase needs to be able to be initialized from voice setMode
and setParam, in addition to the standard initialization through
CreateAudioPatch. This will handle re-initialization of CRS after an
AudioServer restart, as well as with any other instances where CRS
needs to be re-initialized without a new CreateAudioPatch call.

Change-Id: I28e2a2ed280a5b28ea9038e65ae114f854175095
Tags: CRS FR82617