hal: check to avoid metadata update if call is active or mode is in call
During handling of phone mode becoming NORMAL, mode is updated at the
start and call state is updated after voice streams are closed. Incoming
metdata update from framework was relying only on phone mode to update it
to BT. As phone mode was updated to NORMAL, metdata update triggered
reconfiguration for voice streams.

Fixed this by checking both call state as well as phone mode to decide on
whether to ignore metadata update to BT or not.

Change-Id: I30d8a5d89231fd4c068a4305eb793281c9d7b115