hal: Use cached speed during stream open
Speed request can come before the first write as well,
in such cases as pal_stream is not started so because of
invalid pal handle, error is returned to framework implying
speed is not supported, causing a reset in speed to 1.0x in apps.

Don't return error if pal handle is invalid, instead cache the
params and when stream starts, use the cached speed to set
playback rate.

Change-Id: I49c280b1bbb179472e63c5cacb7f7717d4742c09