hal: Update controller during DP device disconnection.
When MST is connected, first DP device gets assigned as AUX_DIGITAL,
second one gets assigned as AUX_DIGITAL1, both have stream/mst index
as 0 and 1 respectively and controller as 0. During disconnection,
controller was not being updated due to which both device disconnection
resulted in disconnection of AUX_DIGITAL1, and AUX_DIGITAL was not getting
disconnected. When SST is connected, previous AUX_DIGITAL was being
used with stream/mst_index as 1 for SST which is not proper resulting in
graph_start failures.

Change-Id: I0da96d996a1f36c8014c347e9d65d705d8e761f6